+++ metamodule/metainstall.sh
@@ -4,12 +4,9 @@
 # Module installation hook
 ############################################
 
-# Constants
-IMG_FILE="/data/adb/metamodule/modules.img"
 META="/data/adb/metamodule"
-MNT_DIR="$META/mnt"
-
 . "$META"/utils.sh || exit 1
+
 unzip -o "$ZIPFILE" module.prop -d "$TMPDIR" >&2
 MODNAME=$(get_prop name "$TMPDIR/module.prop")
 MODID=$(get_prop id "$TMPDIR/module.prop")
@@ -21,7 +18,7 @@ log "Installing module: $MODNAME (ID: $MODID)"
 install_module
 
 if module_requires_overlay_move; then
-    ensure_image_mounted
+    ensure_image_mounted || exit $?
 
     # Run the conflict scan before moving files
     check_conflicts