+++ metamodule/post-mount.sh
@@ -1,4 +1,4 @@
 #!/system/bin/sh
 
-MODDIR="${0%/*}"
-ksud kernel nuke-ext4-sysfs "$MODDIR/mnt"
+# Hide the ext4 backing mount from sysfs using the new external state path.
+ksud kernel nuke-ext4-sysfs "/data/adb/overlayfsx-data/mnt"